Add ability to return needed resource without having to unzip package to extract needed file
It is a pain to have to expect to unzip the entire package to get at the
needed resource. To avoid doing that, added a base class NLPResource that
provides a simple API to get only the needed resource which is loaded via
the classpath.

The second modification is to use the Locale to figure out what language
to use to look-up the associated resource. This required moving the
resource files under the specific language tags.

Added a simple test to ensure that the resources are being loaded and
returned.